Feat: Se agrega pantalla de carga de datos y muestreo para empleados

+++ Seph.Principal.Application/Features/HistorialContratos/Commands/CreateHistorialContratoCommand.cs
@@ -15,7 +15,7 @@ namespace Seph.Principal.Application.Features.HistorialContratos.Commands
         long IdEmpleado,
         long IdInstitucion,
         long IdTipoPersonal,
-        int IdTipoContrato,
+        long IdTipoContrato,
         string? StrOtroTipoContrato,
         long IdArea,
         DateTime DateTimeFechaRegistro,
